Simplify A0Z1 design
Without a crossbar, the "A" above the "Z", on its side, is "<N".

All letters less than (or "left of") "N" have an alphabinary value of
zero, and all other letters have an alphabinary value of one. In other
words, "<N" is the inverse boolean condition of alphabinary value.

I like the double-symbolism of this, and I wasn't happy with the "A"
having a crossbar but no "hole", so I've run with this - in the process,
I'm also removing the "crossbar" from the zero, to keep with the theme
of geometric simplicity.